requirements.txt
langchain
langchain-core
langchain-google-vertexai
langchain-chroma
langchain-text-splitters
google-generativeai
google-auth
google-cloud-aiplatform
chromadb
PyMuPDF
python-pptx
pillow
pydantic
python-dotenv
tqdm
openpyxl
xlrd
email-parser
python-email
numpy
imagehash
aiohttp
